simulation ok

+++ g4l/models/builders/incremental.py
@@ -96,7 +96,7 @@ def count_subsequence_frequencies(df, sample, max_depth, scan_offset):
     #df = df.rename(columns={'index':'node', 0:'freq'})
     df['active'] = 0
     df = create_indexes(df)
-    df2 = sample.F
+    df2 = sample.F.astype(int)
     df2['idx'] = df.reset_index().set_index('node')['node_idx']
     df2 = df2.melt(id_vars=['idx'], var_name='next_symbol', value_name='freq')
     df2 = df2.sort_values(['idx', 'next_symbol']).reset_index(drop=True).set_index('idx')
